refact: delete motorcycle

+++ RentalManager.Application/Commands/Handlers/RemoveMotorcycleCommandHandler.cs
@@ -15,6 +15,9 @@ namespace RentalManager.Application.Commands.Handlers
 
         public async Task Handle(RemoveMotorcycleCommandRequest request)
         {
+            var motorcycle = await _motorcycleRepository.Get(request.MotorcycleId);
+            if (motorcycle == null)
+                throw new Exception("Moto não existe!");
             await _motorcycleRepository.Delete(request.MotorcycleId);
         }
     }
